requirements.txt
streamlit
pandas
tqdm
beautifulsoup4
requests
faiss-cpu
sentence-transformers
transformers
torch
huggingface_hub
scikit-learn
nltk

